+++ resources/views/components/ecommerce/statistics-chart.blade.php
@@ -40,6 +40,7 @@
                         static: true,
                         monthSelectorType: 'static',
                         dateFormat: 'M j',
+                        locale: window.fpConfig ?? { firstDayOfWeek: 1 },
                         defaultDate: [new Date(Date.now() - 6 * 24 * 60 * 60 * 1000), new Date()],
                         prevArrow: '<svg class=\'stroke-current\' width=\'24\' height=\'24\' viewBox=\'0 0 24 24\' fill=\'none\' xmlns=\'http://www.w3.org/2000/svg\'><path d=\'M15.25 6L9 12.25L15.25 18.5\' stroke=\'\' stroke-width=\'1.5\' stroke-linecap=\'round\' stroke-linejoin=\'round\'/></svg>',
                         nextArrow: '<svg class=\'stroke-current\' width=\'24\' height=\'24\' viewBox=\'0 0 24 24\' fill=\'none\' xmlns=\'http://www.w3.org/2000/svg\'><path d=\'M8.75 19L15 12.75L8.75 6.5\' stroke=\'\' stroke-width=\'1.5\' stroke-linecap=\'round\' stroke-linejoin=\'round\'/></svg>',